clang tidy

+++ tasks/golovanov_d_matrix_max_elem/seq/src/ops_seq.cpp
@@ -1,11 +1,9 @@
 #include "golovanov_d_matrix_max_elem//seq/include/ops_seq.hpp"
 
-#include <numeric>
 #include <vector>
 #include <algorithm>
 
 #include "golovanov_d_matrix_max_elem//common/include/common.hpp"
-#include "util/include/util.hpp"
 
 namespace golovanov_d_matrix_max_elem {
 
@@ -18,7 +16,7 @@ GolovanovDMatrixMaxElemSEQ::GolovanovDMatrixMaxElemSEQ(const InType &in) {
 bool GolovanovDMatrixMaxElemSEQ::ValidationImpl() {
   int columns = std::get<0>(GetInput());
   int strokes = std::get<1>(GetInput());
-  return (columns > 0) && (strokes > 0) && (std::get<2>(GetInput()).size() == static_cast<size_t>(strokes * columns)) &&
+  return (columns > 0) && (strokes > 0) && (static_cast<int>(std::get<2>(GetInput()).size()) == (strokes * columns)) &&
          (GetOutput() == 0);
 }
